requirements.txt
requests
asyncio
websockets
discordwebhook
python-dotenv